+++ nslcd/attmap.c
@@ -276,8 +276,8 @@ const char *attmap_get_value(MYLDAP_ENTRY *entry, const char *attr,
       return NULL;
     if (strlen(values[0]) >= buflen)
     {
-      log_log(LOG_ERR, "attmap_get_value(): buffer too small (%d required)",
-              strlen(values[0]));
+      log_log(LOG_ERR, "attmap_get_value(): buffer too small (%lu required)",
+              (unsigned long) strlen(values[0]));
       return NULL;
     }
     strncpy(buffer, values[0], buflen);
